refactor: reorganize views to match new spec
Breaks ALL tests but we will simply have to rewrite them. Many of the endpoints need to be modified to fit new spec, and many are missing and need to be added
