+++ server/api/views/__init__.py
@@ -1,18 +1,13 @@
 from .auth import Login, Register
 from .user import GetUser
-from .permissions import GetPermissions, UpdatePermissions
 from .workspace import (
     WorkspaceView,
-    AddWorkspaceMember,
-    GetWorkspaceMembers,
+    WorkspaceMembersView,
+    WorkspaceShiftsView,
+    WorkspaceRolesView,
 )
-from .role import (
-    CreateRole,
-    GetWorkspaceRoles,
-    AddMemberRole,
-    RemoveMemberRole,
-    DeleteWorkspaceRole,
-    GetMemberRoles,
-    ModifyWorkspaceRole,
-)
-from .shift import CreateShift, ModifyShift, DeleteShift, GetShifts
+
+from .member import MemberView, MemberPermissionsView, MemberRolesView, MemberShiftsView
+from .permissions import PermissionsView
+from .shift import ShiftView, ShiftFilterView
+from .role import RoleView
