+++ server/api/views/workspace.py
@@ -4,7 +4,7 @@ from rest_framework import status
 from rest_framework.permissions import IsAuthenticated
 from rest_framework_simplejwt.authentication import JWTAuthentication
 
-from ..serializers import WorkspaceSerializer
+from ..serializers import WorkspaceSerializer, ShiftSerializer, RoleSerializer
 from ..models import (
     Workspace,
     WorkspaceMember,
@@ -12,6 +12,8 @@ from ..models import (
     MemberPermissions,
     MemberRole,
     WorkspaceRole,
+    Shift,
+    ShiftRequest,
 )
 
 
@@ -271,8 +273,8 @@ class WorkspaceView(APIView):
         return Response(response, status=status.HTTP_200_OK)
 
 
-class AddWorkspaceMember(APIView):
-    """API view for adding a user to a workspace. Requires manage_workspace_members."""
+class WorkspaceMembersView(APIView):
+    """API view managing members of a workspace."""
 
     authentication_classes = [JWTAuthentication]
     permission_classes = [IsAuthenticated]
@@ -346,11 +348,6 @@ class AddWorkspaceMember(APIView):
             response["error"] = "User is already member of this workspace"
             return Response(response, status=status.HTTP_409_CONFLICT)
 
-
-class GetWorkspaceMembers(APIView):
-    authentication_classes = [JWTAuthentication]
-    permission_classes = [IsAuthenticated]
-
     def get(self, request, workspace_id=None):
         """
         workspace_id (required)
@@ -412,4 +409,224 @@ class GetWorkspaceMembers(APIView):
             members_list.append(entry)
 
         response["members"] = members_list
+        return Response(response, status=status.HTTP_200_OK)
+
+
+class WorkspaceShiftsView(APIView):
+    """API view managing shifts of a workspace."""
+
+    authentication_classes = [JWTAuthentication]
+    permission_classes = [IsAuthenticated]
+
+    def post(self, request):
+        """Create a new Shift in the given workspace.
+
+        Requires manage_schedules permission. Accepted body fields:
+        workspace_id (required), role_id (required), start_time (required),
+        end_time (required), member_id (optional).
+
+        :param request: Authenticated HTTP request with shift details in the body.
+        :type request: rest_framework.request.Request
+        :return: Empty success response on creation, or an error response.
+        :rtype: rest_framework.response.Response
+        """
+        response = {"error": {}}
+
+        serializer = ShiftSerializer(data=request.data)
+        if not serializer.is_valid():
+            response["error"]["code"] = 400
+            response["error"][
+                "message"
+            ] = "Invalid request data, start and end time are required."
+            return Response(response, status=status.HTTP_400_BAD_REQUEST)
+
+        # Verify body contains required fields
+        if "role_id" not in request.data:
+            response["error"]["message"] = "Role ID is required."
+            return Response(response, status=status.HTTP_400_BAD_REQUEST)
+        if "workspace_id" not in request.data:
+            response["error"]["message"] = "Workspace ID is required."
+            return Response(response, status=status.HTTP_400_BAD_REQUEST)
+
+        # Verify workspace exists
+        try:
+            workspace = Workspace.objects.get(pk=request.data["workspace_id"])
+        except Workspace.DoesNotExist:
+            response["error"]["message"] = "Workspace does not exist."
+            return Response(response, status=status.HTTP_404_NOT_FOUND)
+
+        # Verify user is part of workspace and has perms to manage schedules
+        try:
+            creator_member = WorkspaceMember.objects.get(
+                user=request.user, workspace=workspace
+            )
+            MemberPermissions.objects.get(member=creator_member, manage_schedules=True)
+        except WorkspaceMember.DoesNotExist:
+            response["error"]["message"] = "You are not a member of this workspace."
+            return Response(response, status=status.HTTP_403_FORBIDDEN)
+        except MemberPermissions.DoesNotExist:
+            response["error"][
+                "message"
+            ] = "You do not have permissions to manage schedules in this workspace."
+            return Response(response, status=status.HTTP_403_FORBIDDEN)
+
+        # Verify role exists and is part of workspace
+        try:
+            role = WorkspaceRole.objects.get(
+                pk=request.data["role_id"], workspace=workspace
+            )
+        except WorkspaceRole.DoesNotExist:
+            response["error"][
+                "message"
+            ] = "Workspace role does not exist or is not part of workspace."
+            return Response(response, status=status.HTTP_404_NOT_FOUND)
+
+        # Verify member is valid if present
+        if "member_id" in request.data:
+            try:
+                member = WorkspaceMember.objects.get(
+                    pk=request.data["member_id"], workspace=workspace
+                )
+            except WorkspaceMember.DoesNotExist:
+                response["error"][
+                    "message"
+                ] = "Member does not exist or is not part of workspace."
+                return Response(response, status=status.HTTP_404_NOT_FOUND)
+
+        # verify dates are valid
+        start_time = serializer.validated_data["start_time"]
+        end_time = serializer.validated_data["end_time"]
+
+        if start_time > end_time:
+            response["error"]["message"] = "Start time cannot be after end time."
+            return Response(
+                response, status=status.HTTP_400_BAD_REQUEST
+            )  # idk if this is the right status code but wtvr
+
+        # could check if start time is before current time if we want to prevent creating shifts in the past, but i think we should allow that since a workplace might want to do that for recordkeeping or smth
+        shift = Shift.objects.create(
+            workspace=workspace,
+            start_time=start_time,
+            end_time=end_time,
+            created_by=creator_member,
+            role=role,
+            open=True,
+        )
+
+        # Add member to shift if present
+        if "member_id" in request.data:
+            shift.member = member
+            shift.open = False
+            shift.save()
+
+        return Response(response, status=status.HTTP_201_CREATED)
+
+    def get(self, request, workspace_id):
+        # TODO
+        return None
+
+
+class WorkspaceRolesView(APIView):
+    """API view managing roles of a workspace."""
+
+    authentication_classes = [JWTAuthentication]
+    permission_classes = [IsAuthenticated]
+
+    def post(self, request):
+        """Create a new WorkspaceRole in the given workspace.
+
+        Requires manage_workspace_roles permission. Accepted body fields:
+        workspace_id (required), name (optional), pay_rate (optional).
+
+        :param request: Authenticated HTTP request with workspace_id and optional
+            role fields in the body.
+        :type request: rest_framework.request.Request
+        :return: Empty success response on creation, or an error response.
+        :rtype: rest_framework.response.Response
+        """
+        response = {"error": {}}
+
+        serializer = RoleSerializer(data=request.data)
+        if not serializer.is_valid():
+            response["error"]["code"] = 400
+            response["error"]["message"] = "Invalid request data"
+            return Response(response, status=status.HTTP_400_BAD_REQUEST)
+
+        # Verify body contains required fields
+        if "workspace_id" not in request.data:
+            response["error"]["message"] = "Workspace ID is required."
+            return Response(response, status=status.HTTP_400_BAD_REQUEST)
+
+        # Verify workspace exists
+        try:
+            workspace = Workspace.objects.get(pk=request.data["workspace_id"])
+        except Workspace.DoesNotExist:
+            response["error"]["message"] = "Workspace does not exist."
+            return Response(response, status=status.HTTP_404_NOT_FOUND)
+
+        # Verify user has permissions to manage workspace roles
+        try:
+            member = WorkspaceMember.objects.get(
+                user=request.user, workspace=request.data["workspace_id"]
+            )
+            MemberPermissions.objects.get(
+                workspace=request.data["workspace_id"],
+                member=member,
+                manage_workspace_roles=True,
+            )
+        except MemberPermissions.DoesNotExist:
+            response["error"][
+                "message"
+            ] = "You do not have permission modify roles in this workspace."
+            return Response(response, status=status.HTTP_403_FORBIDDEN)
+        except WorkspaceMember.DoesNotExist:
+            response["error"]["message"] = "You are not a member of this workspace."
+            return Response(response, status=status.HTTP_403_FORBIDDEN)
+
+        workspace_role = WorkspaceRole.objects.create(workspace=workspace)
+
+        if "name" in serializer.validated_data:
+            workspace_role.name = serializer.validated_data["name"]
+        if "pay_rate" in serializer.validated_data:
+            workspace_role.pay_rate = serializer.validated_data["pay_rate"]
+
+        workspace_role.save()
+
+        return Response(response, status=status.HTTP_201_CREATED)
+
+    def get(self, request):
+        """Return a list of all WorkspaceRoles in the given workspace.
+
+        :param request: Authenticated HTTP request containing workspace_id in the body.
+        :type request: rest_framework.request.Request
+        :return: List of roles with id, name, and pay_rate, or an error response.
+        :rtype: rest_framework.response.Response
+        """
+        response = {"error": {}}
+
+        # Verify body contains required fields
+        if "workspace_id" not in request.data:
+            response["error"]["message"] = "Workspace ID is required."
+            return Response(response, status=status.HTTP_400_BAD_REQUEST)
+
+        # Verify workspace exists
+        try:
+            workspace = Workspace.objects.get(pk=request.data["workspace_id"])
+        except Workspace.DoesNotExist:
+            response["error"]["message"] = "Workspace does not exist."
+            return Response(response, status=status.HTTP_404_NOT_FOUND)
+
+        results = WorkspaceRole.objects.filter(workspace=workspace)
+
+        role_list = [
+            {
+                "id": role.id,
+                "name": role.name,
+                "pay_rate": role.pay_rate,
+            }
+            for role in results
+        ]
+
+        response["roles"] = role_list
+
         return Response(response, status=status.HTTP_200_OK)
